# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aervox｜思隅 API",
    "version": "0.2.0",
    "description": "Turn streaming plus Persona, Anthropic Skills, MCP policy, and GPT-SoVITS contracts."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/v1/skills": {
      "get": {
        "summary": "列出有效 Skills",
        "tags": [
          "Skills"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "X-Workspace-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "X-User-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "X-Actor-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Skills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "skills": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "workspaceId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string",
                            "minLength": 1
                          },
                          "license": {
                            "type": "string"
                          },
                          "compatibility": {
                            "type": "string"
                          },
                          "metadata": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "allowedTools": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "source": {
                            "type": "string",
                            "enum": [
                              "active",
                              "workspace",
                              "imported"
                            ]
                          },
                          "version": {
                            "type": "integer",
                            "exclusiveMinimum": 0
                          },
                          "checksum": {
                            "type": "string",
                            "minLength": 64,
                            "maxLength": 64
                          },
                          "enabled": {
                            "type": "boolean"
                          },
                          "valid": {
                            "type": "boolean"
                          },
                          "validationErrors": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "importedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "workspaceId",
                          "name",
                          "description",
                          "source",
                          "version",
                          "checksum",
                          "enabled",
                          "valid",
                          "validationErrors",
                          "importedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "skills"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/skills/{skillName}": {
      "delete": {
        "summary": "删除工作区 Skill",
        "tags": [
          "Skills"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": true,
            "name": "skillName",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "X-Workspace-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "X-User-Id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "required": false,
            "name": "X-Actor-Id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
...
}